backend/pyproject.toml
[tool.poetry]
name = "pushstart-backend"
version = "0.1.0"
description = "Backend for Pushstart"
authors = ["Gaurav Pooniwala"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
mcp = "^0.1.0"
langchain = "^0.1.0"
langchain-core = "^0.1.0"
langchain-anthropic = "^0.1.0"
langgraph = "^0.0.10"
sqlmodel = "^0.0.14"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
